backend/app/pyproject.toml
[project]
name = "gemini-tales"
version = "0.1.0"
description = "Course Creation Web App"
requires-python = ">=3.10,<3.14"
dependencies = [
    "uvicorn==0.40.0",
    "fastapi==0.123.*",
    "httpx==0.28.*",
    "httpx_sse==0.4.*",
    "google-genai==1.57.*",
    "google-cloud-logging==3.13.0",
    "opentelemetry-exporter-gcp-trace==1.11.0",
    "python-dotenv==1.0.1",
    "Pillow==10.4.*",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
